+++ webstepper/src/SvgDisplay.tsx
@@ -24,6 +24,14 @@ export default function SvgDisplay(props: SvgDisplayPropTypes) {
     const containerRef = useRef<HTMLDivElement>(null);
     const transformRef = useRef<ReactZoomPanPinchContentRef>(null);
     const canvasWidth = props.memoryVizData.configuration?.width ?? 1300;
+    const savedTransform = transformRef.current
+        ? {
+              positionX: transformRef.current.state.positionX,
+              positionY: transformRef.current.state.positionY,
+              scale: transformRef.current.state.scale,
+          }
+        : { positionX: 0, positionY: 0, scale: 1 };
+
     useEffect(() => {
         if (props.memoryVizData.memoryVizInput && containerRef.current) {
             try {
@@ -46,7 +54,12 @@ export default function SvgDisplay(props: SvgDisplayPropTypes) {
                 m.attachInteractivity(svgElement);
 
                 // reset zoom and pan to default when redrawing
-                transformRef.current?.setTransform(0, 0, 1, 0);
+                transformRef.current?.setTransform(
+                    savedTransform.positionX,
+                    savedTransform.positionY,
+                    savedTransform.scale,
+                    0
+                );
             } catch (error) {
                 console.error(error);
             }